+++ packages/server/server.js
@@ -2,20 +2,26 @@ require('dotenv').config();
 var cookieParser = require('cookie-parser');
 var csrf = require('csurf');
 var bodyParser = require('body-parser');
-const axios = require('axios');
 
 const express = require('express');
+const shrinkRay = require('shrink-ray-current');
+
 let app = express();
+
 const expressNunjucks = require('express-nunjucks');
 const fs = require('fs');
 const util = require('util');
 
-const isDev = app.get('env') == 'development';
+const isDev = process.env.APP_ENV != 'production';
 const path = require('path');
 
 app.set('views', __dirname);
 
-const PublicPath = './public';
+const PublicPath =
+  process.env.NODE_ENV == 'production' ? './build' : './public';
+
+// compress all responses
+app.use(shrinkRay());
 
 app.use(bodyParser.json()); // for parsing application/json
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -41,31 +47,6 @@ app.use(function(req, res, next) {
     'Origin, X-Requested-With, Content-Type, Accept, csrf-token, Authorization'
   );
   next();
-});
-
-app.post('/ajax/generateClientSecret', (req, res) => {
-  const url = process.env.API_HOST + '/api/register/device';
-  const deviceId = req.body.browser_device_id;
-  const data = {
-    client_secret: process.env.WEB_API_KEY,
-    device_id: deviceId
-  };
-  axios
-    .post(url, data, {
-      headers: {
-        'Content-Type': 'application/json',
-        'client-secret': process.env.WEB_API_KEY
-      }
-    })
-    .then(value => {
-      res.status(200).json(value.data);
-    })
-    .catch(err => {
-      console.log('ERROR BROH');
-      console.log(err.data);
-      res.status(400);
-    });
-  return;
 });
 
 app.get('*', (req, res) => {
@@ -94,5 +75,5 @@ if (process.env.APP_PORT) {
 }
 
 app.listen(runningPort, () =>
-  console.log('Example app listening on port ' + runningPort + '! ')
+  console.log('Qasir Admin Panel listening on port ' + runningPort + '! ')
 );