+++ webpack.prod.js
@@ -4,28 +4,25 @@ const CleanWebpackPlugin = require('clean-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const ManifestPlugin = require('webpack-manifest-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
+const CompressionPlugin = require('compression-webpack-plugin');
 
 const JSPath = 'assets/js/';
-const localesPath = 'assets/locales';
 
 config.output = {
   filename: JSPath + '[name].[contenthash].js',
   path: path.resolve(__dirname, './build')
 };
-let outputFilename = 'index.html';
-if (process.env.APP_ENV == 'production') {
-  outputFilename = 'index.html';
-}
 
-if (process.env.APP_ENV == 'staging') {
-  outputFilename = 'stage-output.html';
-}
+config.devtool = '';
+
+let outputFilename = 'index.html';
+
 config.plugins.push(
   new CleanWebpackPlugin(['./build/*.js', './build/*.css'], {
     exclude: ['./build/.gitignore', './build/fonts/*']
   }),
   new HtmlWebpackPlugin({
-    title: 'Qasir Mitra Application',
+    title: 'Qasir Application',
     template: './packages/server/index.html',
     minify: true,
     cache: true,
@@ -35,7 +32,7 @@ config.plugins.push(
     fileName: 'manifest.json',
     basePath: process.env.NODE_ENV === 'production' ? '/build/' : '/public/',
     seed: {
-      name: 'App name',
+      name: 'Qasir Application',
       short_name: 'Short app name',
       icons: '',
       start_url: '',
@@ -54,14 +51,29 @@ config.plugins.push(
       from: './packages/assets/locales/',
       to: 'assets/locales/'
     }
-  ])
+  ]),
+  new CompressionPlugin({
+    filename: '[path].gz[query]',
+    algorithm: 'gzip',
+    test: /\.js$|\.css$|\.html$/,
+    threshold: 10240,
+    minRatio: 0.8
+  }),
+  new CompressionPlugin({
+    filename: '[path].br[query]',
+    algorithm: 'brotliCompress',
+    test: /\.(js|css|html|svg)$/,
+    compressionOptions: { level: 11 },
+    threshold: 10240,
+    minRatio: 0.8
+  })
 );
 
 config.optimization = Object.assign(config.optimization, {
   runtimeChunk: 'single',
   splitChunks: {
     cacheGroups: {
-      vendor: {
+      commons: {
         test: /[\\/]node_modules[\\/]/,
         name: 'vendors',
         chunks: 'all'